fix: update_last_revealed
Previously the call to `update_last_revealed` was conditional
on whether the descriptor was present in the changeset, but it
may be the case that the derivation index increases while the
descriptor field of the wallet ChangeSet is None.

We fix this by separating the calls to `insert_descriptor` and
`update_last_revealed` and also changing the signature of
`update_last_revealed` to accept a descriptor id rather than
a keychain.